[FIX] stock: Prevent to reset a 'done' picking in 'draft' state
This change will prevent the document from going from Done to Draft

In some cases a user may make that error functionally. It is only to have an additional validation

In the past, we didn't need this validations because the workflow didn't allow the use of a transition that is not defined on the correct state but since that we don't have workflow so we need manage this inconsistent action from original method in order to avoid a wrong transition.