[FIX] web: handle ´domain´ widget rendering without model
When creating a new record where the ´domain´ widget is set in the view
with an associated model but this model field has no value, an rpc
(search_count) was done on an unknown model, that causes a traceback.

To avoid this, the rpc is not done if the model is not set.