[IMP] project: Remove alias_model field on a project
This field was needed because a rating could be sent to a customer for a task or an issue.

This was buggy, because if a template was set on a shared stage between tasks and issues, the rendering could crash, generate an email with wrong piece of information, and so on.

By removing this field, we can also remove some ugly code that was needed to make that distinction.