[FIX] stock: ACL for product attributes for Inventory managers
If only Inventory is installed,
only the admin could create new product attributes
for products variants. Other managers couldn't.

This revision adds the ACL according to what is done in other modules,
such as in `sale`.

Mostlikely, we should create a product manager group, for which
this rights are given, and make the other groups (sales manager, inventory manager)
inherit from this one.

opw-745735