[FIX] account: avoid duplicated move name
Commits 873bc7afce4 and 83598b55aa0ade attempted to prevent the creation
of two moves with the same name during reconciliation. It works in most
cases, but it is still possible to reconcile the same statement line
with more than one line.

We add an extra check to make sure duplicate never happen.

opw-742018