[FIX] mrp: multi production inconsistencies
When several user run a production on a workorder, things go awry.

For example the end of the "productive" production is computed based on

ProductiveStartTime + DurationExpected - Duration

so if the duration was bigger than the time expected, we easily get a
productive end before the start, and start of the "reduced speed" before
the production actually happened.

eg:

productive time is 30 minutes

user A: start production at 10:00
user B: start production at 10:10

production is set to done at 10:50

=> result before the patch:

user A has been productive from 10:00 to 10:30
user A has not been productive from 10:30 to 10:50
user B has been productive from 10:10 to 10:00
user B has not been productive from 10:00 to 10:50

this is totally erroneous.

=> result after the patch:

user A has been productive from 10:00 to 10:30
user A has not been productive from 10:30 to 10:50
user B has not been productive from 10:10 to 10:50

this is a measure of right, we could argue about wether one or the other
should be productive or not but this is anyway an edge case.

In master we are considering only allowing one user to run a production
at a time but we prefer not to change such a behavior in stable version.

opw-709223
closes #17528